+++ src/cljs/cxengage_javascript_sdk/domain/specs.cljs
@@ -95,6 +95,7 @@
 (s/def ::message string?)
 (s/def ::note-id ::uuid)
 (s/def ::no-session boolean?)
+(s/def ::outbound-ani string?)
 (s/def ::page number?)
 (s/def ::password string?)
 (s/def ::phone-number string?)