+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.26.0] - 2018-11-2
+### Added
+- CXV1-12476 - getUserOutboundIdentifierLists function
+- CXV1-12476 - outboundAni parameter and functionality to the click to dial function
+
 ## [8.25.0] - 2018-11-2
 ### Added
 - Added function to get Data Access User details